feat: moved the admin panel from branch 16, also fixed some bugs

+++ src/main/java/kg/edu/mathbilim/repository/UserTypeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserTypeRepository extends JpaRepository<UserType, Long> {
+public interface UserTypeRepository extends JpaRepository<UserType, Integer> {
     Optional<UserType> findByName(String name);
 
     boolean existsByName(String name);
